Build pattern sequence
The pattern is fixed to 4 digits, can be expressed as an array.
For the first digit, the index of the pattern is modulo position.
For all following digits, we "stretch" the position by dividing, which
will create repetitions.

To create the 1-offset of the pattern, we start the index counting at 1
instead of 0. In the code, we achieve this by first incrementing the
position and then calculating the index.